[Fix] Fixed detail memo

+++ app/src/main/java/jp/gr/java/conf/tmproject/dojoandroid2022/presentation/ui/dialog/warning/DeleteNodeDialogViewModel.kt
@@ -1,4 +1,4 @@
-package jp.gr.java.conf.tmproject.dojoandroid2022.presentation.ui.rodemap.memo.detail
+package jp.gr.java.conf.tmproject.dojoandroid2022.presentation.ui.dialog.warning
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
@@ -7,17 +7,15 @@ import jp.gr.java.conf.tmproject.dojoandroid2022.domain.model.Node
 import jp.gr.java.conf.tmproject.dojoandroid2022.domain.repository.MemoRepository
 import jp.gr.java.conf.tmproject.dojoandroid2022.domain.repository.RoadmapRepository
 import kotlinx.coroutines.flow.MutableSharedFlow
-import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharedFlow
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
-class MemoDetailViewModel @Inject constructor(
+class DeleteNodeDialogViewModel @Inject constructor(
     private val roadmapRepository: RoadmapRepository,
-    private val memoRepository: MemoRepository) : ViewModel() {
-
-    val memo: MutableStateFlow<String> = MutableStateFlow("")
+    private val memoRepository: MemoRepository
+) : ViewModel() {
 
     private val _isDeleteSuccess: MutableSharedFlow<Boolean> = MutableSharedFlow()
     val isDeleteSuccess: SharedFlow<Boolean> = _isDeleteSuccess
@@ -25,13 +23,9 @@ class MemoDetailViewModel @Inject constructor(
     private val _isError: MutableSharedFlow<Boolean> = MutableSharedFlow()
     val isError: SharedFlow<Boolean> = _isError
 
-    fun setSelectedNodeMemo(nodeId: Int) = viewModelScope.launch {
-        memoRepository.loadMemoById(nodeId).forEach {
-            memo.value = it.memo
-        }
-    }
+    fun deleteNode(node: Node?) = viewModelScope.launch {
+        if (node == null) return@launch
 
-    fun deleteNode(node: Node) = viewModelScope.launch {
         runCatching {
             roadmapRepository.deleteNode(node)
             memoRepository.deleteMemo(node.id)
